No runtime errors on missing or wrong video file lists
File lists are given as command line parameters according to
./ravidplay.py -idle: <idle-files> -cntdn: <cntdn-files> -appl: <appl-files>

if -idle: is missing a defined error occurs and the software exits
if -cntdn: is missing a defined error occurs and the software exits
if -appl: is missing the list -idle: will be copied into.

Errors in single files cause a defined error, too. This should be
fixed in this way: An error occurs after ALL files of a list were
checked bad.